refactor(Layout): 合并方向布局文件
将StackLayout和WrapLayout的横向/纵向声明与实现整合到对应主文件中，删除独立H/V文件，并同步更新聚合头和VS工程文件。
