--- sw/src/StackLayoutH.cpp
@@ -1,33 +1,0 @@
-#include "StackLayoutH.h"
-#include "Utils.h"
-#include <cmath>
-
-sw::Size sw::StackLayoutH::MeasureOverride(const Size &availableSize)
-{
-    Size desireSize{};
-    int childCount = this->GetChildLayoutCount();
-
-    for (int i = 0; i < childCount; ++i) {
-        ILayout &item = this->GetChildLayoutAt(i);
-        item.Measure(Size(INFINITY, std::isinf(availableSize.height) ? INFINITY : availableSize.height));
-
-        Size itemDesireSize = item.GetDesireSize();
-        desireSize.width += itemDesireSize.width;
-        desireSize.height = Utils::Max(desireSize.height, itemDesireSize.height);
-    }
-
-    return desireSize;
-}
-
-void sw::StackLayoutH::ArrangeOverride(const Size &finalSize)
-{
-    double width   = 0;
-    int childCount = this->GetChildLayoutCount();
-
-    for (int i = 0; i < childCount; ++i) {
-        ILayout &item       = this->GetChildLayoutAt(i);
-        Size itemDesireSize = item.GetDesireSize();
-        item.Arrange(Rect{width, 0, itemDesireSize.width, finalSize.height});
-        width += itemDesireSize.width;
-    }
-}
