+++ vs/sw.vcxproj
@@ -362,7 +362,6 @@
     <ClInclude Include="..\sw\inc\MenuItem.h" />
     <ClInclude Include="..\sw\inc\MonthCalendar.h" />
     <ClInclude Include="..\sw\inc\MsgBox.h" />
-    <ClInclude Include="..\sw\inc\NotificationObject.h" />
     <ClInclude Include="..\sw\inc\NotifyIcon.h" />
     <ClInclude Include="..\sw\inc\ObservableCollection.h" />
     <ClInclude Include="..\sw\inc\ObservableObject.h" />
@@ -388,8 +387,6 @@
     <ClInclude Include="..\sw\inc\SplitButton.h" />
     <ClInclude Include="..\sw\inc\Splitter.h" />
     <ClInclude Include="..\sw\inc\StackLayout.h" />
-    <ClInclude Include="..\sw\inc\StackLayoutH.h" />
-    <ClInclude Include="..\sw\inc\StackLayoutV.h" />
     <ClInclude Include="..\sw\inc\StackPanel.h" />
     <ClInclude Include="..\sw\inc\StaticControl.h" />
     <ClInclude Include="..\sw\inc\StatusBar.h" />
@@ -410,8 +407,6 @@
     <ClInclude Include="..\sw\inc\WndBase.h" />
     <ClInclude Include="..\sw\inc\WndMsg.h" />
     <ClInclude Include="..\sw\inc\WrapLayout.h" />
-    <ClInclude Include="..\sw\inc\WrapLayoutH.h" />
-    <ClInclude Include="..\sw\inc\WrapLayoutV.h" />
     <ClInclude Include="..\sw\inc\WrapPanel.h" />
   </ItemGroup>
   <ItemGroup>
@@ -479,8 +474,6 @@
     <ClCompile Include="..\sw\src\SplitButton.cpp" />
     <ClCompile Include="..\sw\src\Splitter.cpp" />
     <ClCompile Include="..\sw\src\StackLayout.cpp" />
-    <ClCompile Include="..\sw\src\StackLayoutH.cpp" />
-    <ClCompile Include="..\sw\src\StackLayoutV.cpp" />
     <ClCompile Include="..\sw\src\StackPanel.cpp" />
     <ClCompile Include="..\sw\src\StaticControl.cpp" />
     <ClCompile Include="..\sw\src\StatusBar.cpp" />
@@ -499,8 +492,6 @@
     <ClCompile Include="..\sw\src\Window.cpp" />
     <ClCompile Include="..\sw\src\WndBase.cpp" />
     <ClCompile Include="..\sw\src\WrapLayout.cpp" />
-    <ClCompile Include="..\sw\src\WrapLayoutH.cpp" />
-    <ClCompile Include="..\sw\src\WrapLayoutV.cpp" />
     <ClCompile Include="..\sw\src\WrapPanel.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
